+++ src/services/predicate/lodash_predicates.ts
@@ -23,6 +23,7 @@ import _isString from 'lodash/isString';
 import _isBoolean from 'lodash/isBoolean';
 import _isNumber from 'lodash/isNumber';
 import _isNaN from 'lodash/isNaN';
+import _isObject from 'lodash/isObject';
 
 // wrap the lodash functions to avoid having lodash's TS type definition from being
 // exported, which can conflict with the lodash namespace if other versions are used
@@ -34,3 +35,4 @@ export const isString = (value: any): value is string => _isString(value);
 export const isBoolean = (value: any): value is boolean => _isBoolean(value);
 export const isNumber = (value: any): value is number => _isNumber(value);
 export const isNaN = (value: any) => _isNaN(value);
+export const isObject = (value: any): value is object => _isObject(value);